requirements.txt
django
djangorestframework
django-redis
python-dotenv
psycopg2
